+++ test-parts/part-i-api_frontend/front-main.js
@@ -4,8 +4,8 @@ const cGetListFile = require("./parts/c-get_device_list");
 const cCheckListFile = require("./parts/d-check_device_list");
 const cListAvaliableFile = require("./parts/e-list_avaliable_nodes");
 const cCheckNodeArrayFile = require("./parts/f-check_node_array");
-/*
 const cGetDeviceStatusFile = require("./parts/g-get_device_properties");
+/*
 const cDeleteDevicesFile = require("./parts/h-delete_added_devices");
 const cClearCacheFile = require("./parts/i-clear_cache");
 */
@@ -86,7 +86,7 @@ function callCheckNodeArray()
 		
 		after(function()
 		{
-			//callGetDeviceStatus();
+			callGetDeviceStatus();
 		});
 		
 	});
@@ -100,7 +100,7 @@ function callGetDeviceStatus()
 		
 		after(function()
 		{
-			callDeleteDevices();
+			//callDeleteDevices();
 		});
 		
 	});